+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 ### Latest changes
 
 * VFS-8257 Added closing modal on transition
+* VFS-9257 Fixed ignored eslint issues
 * VFS-10875 Handling large amount of storages and spaces in Oneprovider panel
 * VFS-10563 Add record info with icon component to space table
 * VFS-10886 Replaced browser service usages with Browser util